# docs/05-api-openapi.yaml
openapi: 3.0.3
info:
  title: CareerAI API
  description: |
    API REST de CareerAI (Next.js App Router).
    Authentification : session Supabase via cookies (navigateur) ou cookie de session après login.
    Base URL production : https://www.careerai.live
  version: 1.0.0
  contact:
    name: CareerAI
    url: https://www.careerai.live
servers:
  - url: https://www.careerai.live
    description: Production
  - url: http://localhost:3000
    description: Développement local
...
paths:
...
  /api/recruiter/quizzes/{id}/send:
    post:
      tags: [recruiter]
      summary: Envoyer un quiz par email à un candidat
      security:
        - cookieAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [candidateId, candidateEmail]
              properties:
                candidateId:
                  type: string
                  format: uuid
                candidateEmail:
                  type: string
                  format: email
      responses:
        '200':
          description: Quiz envoyé
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  quizLink:
                    type: string
                  sentToRecruiterFallback:
                    type: boolean
        '400':
          description: Quiz inactif ou données manquantes
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Session Supabase (cookies après login). Voir documentation Supabase Auth.
  responses:
    Unauthorized:
      description: Non autorisé — session absente ou expirée
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
...